Rename Hangouts Chat to Google Chat
Add/keep `hangouts` as an alias scheme for `googlechat` for backward
compatibility.
