+++ pkg/router/servicemap.go
@@ -3,8 +3,8 @@ package router
 import (
 	"github.com/containrrr/shoutrrr/pkg/services/discord"
 	"github.com/containrrr/shoutrrr/pkg/services/generic"
+	"github.com/containrrr/shoutrrr/pkg/services/googlechat"
 	"github.com/containrrr/shoutrrr/pkg/services/gotify"
-	"github.com/containrrr/shoutrrr/pkg/services/hangouts"
 	"github.com/containrrr/shoutrrr/pkg/services/ifttt"
 	"github.com/containrrr/shoutrrr/pkg/services/join"
 	"github.com/containrrr/shoutrrr/pkg/services/logger"
@@ -27,7 +27,8 @@ var serviceMap = map[string]func() t.Service{
 	"discord":    func() t.Service { return &discord.Service{} },
   "generic":    func() t.Service { return &generic.Service{} },
 	"gotify":     func() t.Service { return &gotify.Service{} },
-	"hangouts":   func() t.Service { return &hangouts.Service{} },
+	"googlechat": func() t.Service { return &googlechat.Service{} },
+	"hangouts":   func() t.Service { return &googlechat.Service{} },
 	"ifttt":      func() t.Service { return &ifttt.Service{} },
 	"join":       func() t.Service { return &join.Service{} },
 	"logger":     func() t.Service { return &logger.Service{} },
