+++ pkg/services/googlechat/googlechat_json.go
@@ -1,0 +1,6 @@
+package googlechat
+
+// JSON is the actual payload being sent to the Google Chat API.
+type JSON struct {
+	Text string `json:"text"`
+}
